requirements.txt
amqp==5.0.6
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.19
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
packaging==20.9
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
PyJWT==2.0.1
pyparsing==2.4.7
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
